+++ lib/src/utils/hex.ts
@@ -4,12 +4,43 @@
 /**
  * Hex utilities
  *
- * Re-exports hex conversion functions from key-derivation
+ * Canonical location for hex conversion functions
  */
 
 import type { Address } from "../schemas"
 
-export { hexToUint8Array, uint8ArrayToHex } from "./key-derivation"
+/**
+ * Convert a hex string to Uint8Array
+ *
+ * @param hexString - Hex string (e.g., "deadbeef")
+ * @returns Uint8Array
+ */
+export function hexToUint8Array(hexString: string): Uint8Array {
+  // Remove any whitespace and ensure even length
+  const hex = hexString.replace(/\s/g, "")
+  if (hex.length % 2 !== 0) {
+    throw new Error("Invalid hex string: length must be even")
+  }
+
+  const bytes = new Uint8Array(hex.length / 2)
+  for (let i = 0; i < hex.length; i += 2) {
+    bytes[i / 2] = parseInt(hex.substring(i, i + 2), 16)
+  }
+
+  return bytes
+}
+
+/**
+ * Convert a Uint8Array to hex string
+ *
+ * @param bytes - Uint8Array to convert
+ * @returns Hex string (e.g., "deadbeef")
+ */
+export function uint8ArrayToHex(bytes: Uint8Array): string {
+  return Array.from(bytes)
+    .map((b) => b.toString(16).padStart(2, "0"))
+    .join("")
+}
 
 /**
  * Create a validated hex address string (40 lowercase hex chars).
